move mangle/unmangle to separate file
